refactor(content): remove content pkg barrel files
