+++ packages/content/src/__tests__/content-plugin.test.ts
@@ -2,11 +2,7 @@ import { createMockLogger } from "@bluecadet/launchpad-testing/test-utils.ts";
 import { PluginDriver } from "@bluecadet/launchpad-utils/plugin-driver";
 import { describe, expect, it, vi } from "vitest";
 import { contentConfigSchema } from "../content-config.js";
-import {
-	ContentError,
-	ContentPluginDriver,
-	defineContentPlugin,
-} from "../content-plugin-driver.js";
+import { ContentError, ContentPluginDriver, defineContentPlugin } from "../content-plugin.js";
 import { DataStore } from "../utils/data-store.js";
 
 describe("ContentPluginDriver", () => {
