+++ packages/content/src/content-plugin.ts
