+++ packages/content/src/fetching/__tests__/fetch-stages.test.ts
@@ -3,9 +3,9 @@ import { vol } from "memfs";
 import { errAsync, okAsync } from "neverthrow";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import type { ResolvedContentConfig } from "../../content-config.js";
-import type { ContentPluginDriver } from "../../content-plugin-driver.js";
-import { ContentError } from "../../content-plugin-driver.js";
-import { defineSource } from "../../sources/source.js";
+import type { ContentPluginDriver } from "../../content-plugin.js";
+import { ContentError } from "../../content-plugin.js";
+import { defineSource } from "../../source.js";
 import type { DataStore } from "../../utils/data-store.js";
 import type { FetchStageContext } from "../fetch-context.js";
 import {
