+++ packages/content/src/fetching/fetch-stages.ts
@@ -7,8 +7,8 @@
 
 import chalk from "chalk";
 import { errAsync, okAsync, ResultAsync } from "neverthrow";
-import { ContentError } from "../content-plugin-driver.js";
-import type { ContentSource } from "../sources/source.js";
+import { ContentError } from "../content-plugin.js";
+import type { ContentSource } from "../source.js";
 import { FetchLogger } from "../utils/fetch-logger.js";
 import * as FileUtils from "../utils/file-utils.js";
 import type { FetchStageContext } from "./fetch-context.js";
