--- packages/content/src/index.ts
@@ -1,13 +1,0 @@
-import LaunchpadContent from "./launchpad-content.js";
-
-export * from "./content-commands.js";
-export * from "./content-config.js";
-export * from "./content-events.js";
-export * from "./content-plugin-driver.js";
-export * from "./content-state.js";
-export * from "./launchpad-content.js";
-export * from "./plugins/index.js";
-export * from "./sources/index.js";
-export * from "./utils/file-utils.js";
-
-export default LaunchpadContent;
