+++ packages/content/src/sources/contentful-source.ts
@@ -1,7 +1,7 @@
 import type { Asset, Entry } from "contentful";
 import { z } from "zod";
+import { defineSource } from "../source.js";
 import { fetchPaginated } from "../utils/fetch-paginated.js";
-import { defineSource } from "./source.js";
 
 // If deliveryToken is provided, then previewToken is optional.
 const contentfulCredentialsSchema = z.union(
