--- packages/content/src/sources/index.ts
@@ -1,6 +1,0 @@
-export { default as airtableSource } from "./airtable-source.js";
-export { default as contentfulSource } from "./contentful-source.js";
-export { default as jsonSource } from "./json-source.js";
-export { default as sanitySource } from "./sanity-source.js";
-export * from "./source.js";
-export { default as strapiSource } from "./strapi-source.js";
